XEngine-V7.46.0.1001 Build20221228
其他更新
    安装脚本现在支持ROCKYLINUX了
    安装脚本修正Centos无法正确安装rb_libtorrent库的问题
    安装脚本删除MACOS下安装xcode的功能,需要用户自己安装
    安装脚本现在对MACOS的BREW命令会判断是否安装了
    应用程序现在不在依赖任何模块了
平台更新
    ANDROID改造已经支持一下组件模块
        基础组件(基础模块,算法魔法)
        核心组件(池管理模块,核心服务模块,协议模块,核心函数模块,黑白名单模块)
        客户端组件(客户端模块)
        系统组件(系统函数模块,profile文件处理模块)
        帮助组件(授权模块,文件打包解包模块,网络组包解包模块,日志模块)
        网络组件(压力测试模块)
        标准组件(MQTT协议模块,NAT协议模块,代理协议模块,SDP协议模块,SIP协议模块,SNMP协议模块,SSDP协议模块)
        流媒体组件(RTCP协议模块,RTP协议模块)
    MACOS平台编译修正一处警告问题
客户端组件
    客户端模块
        SCTP客户端已经被移除.兼容性不好
标准协议组件
    UPNP协议模块-删除
    SSDP协议模块-增加
        协议打包增加了设备搜索请求协议,支持ST和USERAGENT以及IP版本设置
        协议打包增加了设备搜索回复协议,支持ST和USN和LOCATION URL以及版本和时间设置
        协议打包增加了通知协议,支持NT,USN,LOCATION URL,超时时间和IP版本设置
        协议打包增加了离开协议,支持USN和IP版本设置
        协议解析增加了解析SSDP协议数据为数据结构体的功能函数,支持导出头字段
        协议解析增加了获取指定头字段对应指定内容值的功能函数
        协议帮助增加了UPNP打包协议支持,支持打包UPNP操作协议(IP地址,获取映射,添加映射,删除映射)
核心组件
    核心服务模块
        多播通信组播和广播现在套接字重用使用已有代码了
        套接字组播通信增加了新的组播使用网卡方法
        SCTP服务已经被移除,因为只有LINUX支持,兼容性不好,以后有机会添加
基础组件
    基础模块
        句柄生成类现在不在使用guid库了,使用新的标准库方式生成GUID

+++ XEngine_LINEnv.sh
@@ -10,8 +10,8 @@ m_CMDBrew=0
 m_EvnFileClear=0
 m_EnvAuthBreak=0
 m_EnvRelease=0
-m_EnvRPM='git redhat-lsb libuuid openssl-libs libcurl mariadb-connector-c zlib minizip ffmpeg-libs lksctp-tools SDL2 mongo-c-driver-libs libpq libsqlite3x libnghttp2 rb_libtorrent'
-m_EnvAPT='git lsb-core lsb-release libuuid1 libssl1.1 libcurl4 libmysqlclient21 zlib1g libminizip1 libsctp1 libsdl2-2.0-0 libbson-1.0-0 libmongoc-1.0-0 libpq5 libsqlite3-0 libnghttp2-14 libavcodec58 libavdevice58 libavfilter7 libavformat58 libpostproc55 libswresample3 libswscale5 libtorrent-rasterbar9'
+m_EnvRPM='git redhat-lsb openssl-libs libcurl mariadb-connector-c zlib minizip ffmpeg-libs SDL2 mongo-c-driver-libs libpq libsqlite3x libnghttp2 rb_libtorrent'
+m_EnvAPT='git lsb-core lsb-release libssl1.1 libcurl4 libmysqlclient21 zlib1g libminizip1 libsdl2-2.0-0 libbson-1.0-0 libmongoc-1.0-0 libpq5 libsqlite3-0 libnghttp2-14 libavcodec58 libavdevice58 libavfilter7 libavformat58 libpostproc55 libswresample3 libswscale5 libtorrent-rasterbar9'
 m_EnvMAC='curl openssl1.1 sqlite zlib minizip mongo-c-driver mysql-client libpq libtorrent-rasterbar libnghttp2 ffmpeg@4'
 
 #打印环境
@@ -20,7 +20,7 @@ function InstallEnv_Print()
 	echo -e "\033[32m|***************************************************************************|\033[0m"
 	echo -e "\033[33m                 XEngine-Toolkit Linux和Mac版本环境安装脚本                    \033[0m"
 	echo -e "\033[33m                       运行环境：Linux x64 AND MacOS x64                      \033[0m"
-	echo -e "\033[33m                       脚本版本：Ver 7.40.0.1001                              \033[0m"
+	echo -e "\033[33m                       脚本版本：Ver 7.46.0.1001                              \033[0m"
 	echo -e "\033[33m                  安装环境的时候请检查所有三方库下载安装成功                     \033[0m"
 	echo -e "\033[32m|***************************************************************************|\033[0m"
 	echo -e "\033[44;37m当前时间：$m_EnvTimer 执行用户：$m_EnvExecName 你的环境：$m_EnvCurrent\033[0m"
@@ -36,6 +36,9 @@ function InstallEnv_CheckEnv()
 	if [ $(uname) == "Linux" ] ; then
 		m_EnvRelease=0
 		if grep -Eqi "CentOS" /etc/issue || grep -Eq "CentOS" /etc/*-release; then
+			m_EnvRelease=1
+			m_EnvCurrent=$(cat /etc/redhat-release)
+		elif grep -Eqi "Rocky Linux" /etc/issue || grep -Eq "Rocky Linux" /etc/*-release; then
 			m_EnvRelease=1
 			m_EnvCurrent=$(cat /etc/redhat-release)
 		elif grep -Eqi "Red Hat Enterprise Linux Server" /etc/issue || grep -Eq "Red Hat Enterprise Linux Server" /etc/*-release; then
@@ -113,8 +116,9 @@ function InstallEnv_Checkepel()
 			if test -z `rpm -qa | grep $rpmfusion`
 			then 
 				echo -e "\033[35m不存在rpmfusion扩展源，将开始安装。。。\033[0m"
-				sudo dnf install --nogpgcheck https://mirrors.rpmfusion.org/free/el/rpmfusion-free-release-8.noarch.rpm https://mirrors.rpmfusion.org/nonfree/el/rpmfusion-nonfree-release-8.noarch.rpm -y
-				sudo dnf config-manager --set-enabled powertools
+				sudo dnf install --nogpgcheck https://mirrors.rpmfusion.org/free/el/rpmfusion-free-release-$(rpm -E %rhel).noarch.rpm https://mirrors.rpmfusion.org/nonfree/el/rpmfusion-nonfree-release-$(rpm -E %rhel).noarch.rpm
+				sudo dnf config-manager --enable powertools
+				sudo dnf module enable mariadb-devel -y
 				echo -e "\033[41;33m$rpmfusion 安装完毕\033[0m"
 			else
 				echo -e "\033[41;37mrpmfusion 扩展源存在。。。\033[0m"
@@ -130,19 +134,23 @@ function InstallEnv_Checkepel()
 		fi
 	elif [ "$m_EnvRelease" -eq "3" ] ; then 
 		if [ "$m_CMDBrew" -eq "1" ] ; then
-			echo -e "\033[31mBrew配置为由系统安装。。。\033[0m"
-			xcode-select --install
-			export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"
-			export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"
-			export HOMEBREW_BOTTLE_DOMAIN="https://mirrors.tuna.tsinghua.edu.cn/homebrew-bottles"
-			git clone --depth=1 https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/install.git brew-install
-			/bin/bash brew-install/install.sh
-			rm -rf brew-install
+			echo -e "\033[31mMacos检查是否安装XCode。。。\033[0m"
+			if ! type brew >/dev/null 2>&1; then
+   				echo '\033[31mbrew 未安装,开始安装brew。。。\033[0m';
+				export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"
+				export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"
+				export HOMEBREW_BOTTLE_DOMAIN="https://mirrors.tuna.tsinghua.edu.cn/homebrew-bottles"
+				git clone --depth=1 https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/install.git brew-install
+				/bin/bash brew-install/install.sh
+				rm -rf brew-install
 		
-			echo 'export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"' >> /Users/$m_EnvExecName/.zprofile
-    		echo 'export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"' >> /Users/$m_EnvExecName/.zprofile
-			export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"
-    		export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"
+				echo 'export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"' >> /Users/$m_EnvExecName/.zprofile
+    			echo 'export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"' >> /Users/$m_EnvExecName/.zprofile
+				export HOMEBREW_BREW_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/brew.git"
+    			export HOMEBREW_CORE_GIT_REMOTE="https://mirrors.tuna.tsinghua.edu.cn/git/homebrew/homebrew-core.git"
+			else
+				echo '\033[31mbrew 已安装\033[0m';
+			fi
 		else
 			echo -e "\033[31mBrew配置为用户自己安装。。。\033[0m"
 		fi
